refactor: concert-card 컴포넌트 다른 도메인에서 활용하기 위해 폴더 경로 이동 #124
