+++ src/shared/components/features/concert/concert-card/concert-card.module.scss
